Added test spec suit with a sample test using Jasmine,Karma & Browserify * Browserify is required in order to use require() in test for running them in browsers.
Browserify your test code before sending it to browser if you want to use stuff like 'require()' function. Browserify will provide require() for browser.

To browserify, do the following:

 1. npm install browserify --save-dev
 2. Add browserify to your framework field in karma.conf.js
     frameworks: ['jasmine','browserify']
 3. Add browserify to your plugin field in karma.conf.js
     plugins : [
      'karma-chrome-launcher',
      'karma-jasmine',
      'karma-browserify'
    ]
 4.Add browserify to your pre processor field in karma.conf.js
 preprocessors: {
      'src/spec/*.spec.js':['browserify']
    },

Step#4 is v.imp because this is the one which actually puts require() for browsers before sending the test to browsers.